requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
